+++ tests/assets/testpackage/testpackage/high/blue/one.py
@@ -1,1 +1,2 @@
 from testpackage import utils
+from testpackage.indirect import yellow